Update Stonks.js
- now uses parseFloat instead of parseInt

+++ js/stonks.js
@@ -12,14 +12,14 @@ function stonk(){
 	
 	var arr = $('form[name="cForm"]').serializeArray();
 
-	var x11 = parseInt(arr[0].value);
-	var x12 = parseInt(arr[1].value);
+	var x11 = parseFloat(arr[0].value);
+	var x12 = parseFloat(arr[1].value);
 	
-	var x21 = parseInt(arr[2].value);
-	var x22 = parseInt(arr[3].value);
+	var x21 = parseFloat(arr[2].value);
+	var x22 = parseFloat(arr[3].value);
 	
-	var x31 = parseInt(arr[4].value);
-	var x32 = parseInt(arr[5].value);
+	var x31 = parseFloat(arr[4].value);
+	var x32 = parseFloat(arr[5].value);
 	
 	var investment = x11 + x21 + x31;
 	var worth = x12 + x22 + x32;
